Abstracted out contrib amounts into shared component.
